pyproject.toml
[project]
name = "systech-aidd"
version = "0.1.0"
description = "AI-powered Telegram bot assistant"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "pydantic-settings>=2.0.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "DTZ", "PIE", "RET", "SIM", "ARG"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
# Note: --cov will be added in iteration 6 when tests are implemented

